Update and rename a2 to Google_Placement_Question
You have two inputs 
I1: No. of rows; I2: Amount of wine 
Row no. indicates no. of glasses in that row and each glass have a capacity of 1 unit after that equal amount of wine will be inside glasses of next row if all the glasses are full then wine will be on a tray
Output:
amount of wine in each glasses and the tray